+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Full dataset added to [our Hugging Face Hub](https://huggingface.co/datasets/MICS-Lab/novae)
 - Official Docker images released - [see here](https://hub.docker.com/r/quentinblampey/novae)
 - Support `reference=None` in `fine_tune` (random initialization) and used as default
+- Added `fast` argument in `novae.plot.domains` to have a quick (but less accurate) rendering of domains.
 
 ### Changed
 - Update hf-hub to use `xet` for faster dataset download from Hugging Face Hub